Work towards [ 1382109 ] new alreadyIncluded: batched fingerprint merge * DiskFPMergeUriUniqFilter.java     check for null on initial (no-disk-FPs-yet) pass

+++ heritrix/src/java/org/archive/crawler/util/DiskFPMergeUriUniqFilter.java
@@ -110,7 +110,9 @@ public class DiskFPMergeUriUniqFilter extends FPMergeUriUniqFilter {
             if(oldFps!=null) {
                 oldFps.close();
             }
-            oldFpsFile.delete();
+            if(oldFpsFile!=null) {
+                oldFpsFile.delete();
+            }
         } catch (IOException e) {
             throw new RuntimeException(e);
         }
